Modified product to have double values instead of floating types.

+++ src/main/java/com/bridgephase/store/Product.java
@@ -3,11 +3,11 @@ package com.bridgephase.store;
 public class Product {
 	private String upc;
 	private String name;
-	private Float wholesalePrice;
-	private Float retailPrice;
+	private double wholesalePrice;
+	private double retailPrice;
 	private Integer quantity;
 
-	public Product(String upc, String name, float wholesalePrice, float retailPrice, int quantity) {
+	public Product(String upc, String name, double wholesalePrice, double retailPrice, int quantity) {
 		this.upc = upc;
 		this.name = name;
 		this.wholesalePrice = wholesalePrice;
